rework page structure to prioritise who is that pokemon
