+++ pages/index.js
@@ -1,17 +1,15 @@
 import Head from "next/head";
-import styles from "../styles/Home.module.css";
-import PokedexSearch from "../components/PokedexSearch/PokedexSearch";
-import Link from "next/link";
+import styles from "../styles/whosThatPokemon.module.css";
+import WhoIsThatPokemon from "../components/WhoIsThatPokemon/WhoIsThatPokemon";
 
-export default function Home() {
+export default function whosThatPokemon() {
   return (
     <div className={styles.container}>
       <Head>
-        <title>Ye Olde Pokemon</title>
+        <title>Who is that Pokemon?</title>
         <link rel="icon" href="/favicon.png" />
       </Head>
-      <Link href="/whosthatpokemon">Try and guess the pokemon!</Link>
-        <PokedexSearch />
+        <WhoIsThatPokemon />
     </div>
   );
 }
